refactor Context lifecycle
The new context is a component and uses the component lifecycle of Unity
to prevent resources from being leaked when a new scene is loaded
and to allow multiple contexts to coexist.

Furthermore, setup and integrity logic has been moved to a separate
static class and is only done once at startup.